// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io

export interface User {
    user:     string;
    password: string;
}

export interface LoginResponse {
    status:     boolean;
    codeStatus: string;
    message:    string;
    data:       string;
}

// Generated by https://quicktype.io

export interface DataResponse {
    status:     boolean;
    codeStatus: string;
    message:    string;
    data:       string;
}